position pk 값 변경에 맞게 코드 수정

+++ src/main/java/kr/co/shoulf/web/control/PositionController.java
@@ -17,7 +17,7 @@ public class PositionController {
 
     @GetMapping(value = "/{positionNo}/detail", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public List<PositionDetail> getPositionDetailList(@PathVariable("positionNo") Long no) {
+    public List<PositionDetail> getPositionDetailList(@PathVariable("positionNo") Integer no) {
         return positionDetailService.readPositionDetail(no);
     }
 }
